added xerces_string_janitor class
Ordinarily you'd expect to be able to use Xerces' Array janitor class.  However,
Windows exes and dlls use seperate heaps, so you can't delete something in one
module that was allocated by another.  It's a pita.  The solution here is to useXMLString::release.
Although this simply does a delete[], the deletion occurs in the Xerces DLL, so
it will work correctly.